FEA Add cupy_utils with rmm cupy ary
