+++ python/cuml/utils/numba_utils.py
@@ -18,6 +18,7 @@ import cupy as cp
 import numpy as np
 import rmm
 
+from cuml.utils.cupy_utils import rmm_cupy_ary
 from cuml.utils.import_utils import check_min_cupy_version
 from numba import cuda
 from numba.cuda.cudadrv.driver import driver
@@ -36,30 +37,18 @@ def row_matrix(df):
 
     # using_allocator was introduced in CuPy 7. Once 7+ is required,
     # this check can be removed alongside the alternative code path.
-    if check_min_cupy_version("7.0"):
-        with cp.cuda.memory.using_allocator(rmm.rmm_cupy_allocator):
-            row_major = cp.array(col_major, order='C')
+    # if check_min_cupy_version("7.0"):
+    #     with cp.cuda.memory.using_allocator(rmm.rmm_cupy_allocator):
+    #         row_major = cp.array(col_major, order='C')
 
-    else:
-        temp_row_major = cp.array(col_major, order='C')
-        row_major = _rmm_cupy6_array_like(temp_row_major, order='C')
-        cp.copyto(row_major, temp_row_major)
+    # else:
+    #     temp_row_major = cp.array(col_major, order='C')
+    #     row_major = _rmm_cupy6_array_like(temp_row_major, order='C')
+    #     cp.copyto(row_major, temp_row_major)
+
+    row_major = rmm_cupy_ary(cp.array, col_major, order='C')
 
     return cuda.as_cuda_array(row_major)
-
-
-def _rmm_cupy6_array_like(ary, order):
-    nbytes = np.ndarray(ary.shape,
-                        dtype=ary.dtype,
-                        strides=ary.strides,
-                        order=order).nbytes
-    memptr = cp.cuda.MemoryPointer(rmm.rmm.RMMCuPyMemory(nbytes), 0)
-    arr = cp.ndarray(ary.shape,
-                     dtype=ary.dtype,
-                     memptr=memptr,
-                     strides=ary.strides,
-                     order=order)
-    return arr
 
 
 @cuda.jit
